+++ packages/gh/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for `gh`
 
+## Unreleased
+
+- Add `GH.Auth.Status` module for checking GitHub CLI authentication status
+
 ## 0.2
 
 - Fix `-f` option